Removed useless _generic.py, fixed(hopefully) the examples
