+++ nengin/examples/flappy.py
@@ -1,5 +1,6 @@
-import nengin as ng
-from nengin import Scene, addScene, screen, Vector
+import nengin.ng as ng
+from nengin.ng import screen
+from nengin import Scene, addScene, Vector
 from pygame import FRect as Rect, K_SPACE
