+++ nengin/examples/jumper.py
@@ -1,6 +1,7 @@
 from pygame.key import ScancodeWrapper
-import nengin as ng
-from nengin import Scene, addScene, screen, Vector
+import nengin.ng as ng
+from nengin.ng import screen
+from nengin import Scene, addScene, Vector
 from pygame import font, FRect as Rect, K_SPACE
 from pygame._sdl2.video import Texture			# pyright: ignore
 from random import randint
