+++ nengin/glng.py
@@ -3,7 +3,7 @@ class GLNenginError(Exception): pass
 if __name__ == "__main__":
 	raise GLNenginError("Run Your own script. Not GlNengin!!!!")
 
-from ._generic import (window,GenericScene,GenericGame,Vector,add_scene,addScene,CLOCK)
+from . import (window,GenericScene,GenericGame,Vector,add_scene,addScene,CLOCK)
 import pygame as pg
 from OpenGL.GL import (
 	glClearColor, glClear, GL_COLOR_BUFFER_BIT,glBegin,GL_LINE_LOOP,
